Use np.int8 instead of np.int64 for action mask dtype
This is more memory sufficient for only 0-1's.
And seems to be the standard for action mask at least for ray.rllib,
as shown in `action_mask_key` documentation at
https://docs.ray.io/en/latest/rllib/rllib-training.html

+++ skdecide/builders/domain/events.py
@@ -384,7 +384,7 @@ class Events:
                     1 if applicable_actions.contains(a) else 0
                     for a in action_space.get_elements()
                 ],
-                dtype=np.int64,
+                dtype=np.int8,
             )
         else:
             # multi agent
@@ -394,7 +394,7 @@ class Events:
                         1 if agent_applicable_actions.contains(a) else 0
                         for a in action_space[agent].get_elements()
                     ],
-                    dtype=np.int64,
+                    dtype=np.int8,
                 )
                 for agent, agent_applicable_actions in applicable_actions.items()
             }
